auto

+++ src/arch/fvit_spatial.py
@@ -32,7 +32,7 @@ class EncoderBlock(nn.Module):
         fourier_dims = hidden_dim // 2
         mlp_input_dims = hidden_dim + fourier_dims
 
-        self.nin_conv = nn.Conv1d(fourier_dims+1, fourier_dims-1, 1)
+        self.nin_conv = nn.Conv1d(fourier_dims+1, fourier_dims-1, 1, dtype=torch.cfloat)
 
         self.fourier_attention = nn.MultiheadAttention(fourier_dims, fourier_dims//2, dropout=attention_dropout, batch_first=True)
 
@@ -49,9 +49,7 @@ class EncoderBlock(nn.Module):
         a = a + input
 
         f = torch.fft.rfft2(x, norm='ortho')
-
         f = self.nin_conv(f)
-
         f, _ = self.fourier_attention(f, f, f, need_weights=False)
         f = torch.fft.irfft2(x, norm='ortho')
         f = self.dropout(f)
